requirements.txt
beautifulsoup4
langchain
langchain-core
langchain-community
langchain-openai
more-itertools
pandas
pymongo
dateparser
python-dotenv
requests
scrapy